fix: clippy -D warnings and SessionMaterialization callers (v0.1.80)
Rust 1.93 pre-commit/CI: plasm-core/runtime/agent-core hygiene; plasm-cli,
plasm-repl, and evm_demo use SessionMaterialization for execute().

+++ crates/plasm-runtime/src/http_resilience.rs
@@ -69,7 +69,7 @@ impl ResilientHttpTransport {
         let exp = self
             .policy
             .initial_backoff
-            .saturating_mul(2u32.saturating_pow(attempt.saturating_sub(1)) as u32);
+            .saturating_mul(2u32.saturating_pow(attempt.saturating_sub(1)));
         let base = exp.min(self.policy.max_backoff);
         let base = retry_after.map(|r| r.max(base)).unwrap_or(base);
         jitter_duration(base, url, attempt)
